Move quick action settings into own subsection
