+++ osu.Game/Localisation/GeneralSettingsStrings.cs
@@ -30,6 +30,11 @@ namespace osu.Game.Localisation
         public static LocalisableString Prefer24HourTimeDisplay => new TranslatableString(getKey(@"prefer_24_hour_time_display"), @"Prefer 24-hour time display");
 
         /// <summary>
+        /// "Quick Actions"
+        /// </summary>
+        public static LocalisableString QuickActionsHeader => new TranslatableString(getKey(@"quick_actions_header"), @"Quick Actions");
+
+        /// <summary>
         /// "Updates"
         /// </summary>
         public static LocalisableString UpdateHeader => new TranslatableString(getKey(@"update_header"), @"Updates");
