+++ osu.Game/Overlays/Settings/Sections/GeneralSection.cs
@@ -7,19 +7,12 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Localisation;
 using osu.Game.Graphics;
 using osu.Game.Localisation;
-using osu.Game.Online.Chat;
 using osu.Game.Overlays.Settings.Sections.General;
 
 namespace osu.Game.Overlays.Settings.Sections
 {
     public partial class GeneralSection : SettingsSection
     {
-        [Resolved(CanBeNull = true)]
-        private FirstRunSetupOverlay? firstRunSetupOverlay { get; set; }
-
-        [Resolved(CanBeNull = true)]
-        private OsuGame? game { get; set; }
-
         public override LocalisableString Header => CommonStrings.General;
 
         public override Drawable CreateIcon() => new SpriteIcon
@@ -28,33 +21,13 @@ namespace osu.Game.Overlays.Settings.Sections
         };
 
         [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
+        private void load()
         {
             Children = new Drawable[]
             {
-                new SettingsButton
-                {
-                    Text = GeneralSettingsStrings.RunSetupWizard,
-                    Keywords = new[] { @"first run", @"initial", @"getting started", @"import", @"tutorial", @"recommended beatmaps" },
-                    TooltipText = FirstRunSetupOverlayStrings.FirstRunSetupDescription,
-                    Action = () => firstRunSetupOverlay?.Show(),
-                },
-                new SettingsButton
-                {
-                    Text = GeneralSettingsStrings.LearnMoreAboutLazer,
-                    TooltipText = GeneralSettingsStrings.LearnMoreAboutLazerTooltip,
-                    BackgroundColour = colours.YellowDark,
-                    Action = () => game?.ShowWiki(@"Help_centre/Upgrading_to_lazer")
-                },
-                new SettingsButton
-                {
-                    Text = GeneralSettingsStrings.ReportIssue,
-                    TooltipText = GeneralSettingsStrings.ReportIssueTooltip,
-                    BackgroundColour = colours.DarkOrange2,
-                    Action = () => game?.OpenUrlExternally(@"https://osu.ppy.sh/community/forums/topics/create?forum_id=5", LinkWarnMode.NeverWarn)
-                },
                 new LanguageSettings(),
                 new UpdateSettings(),
+                new QuickActionSettings(),
             };
         }
     }
